PIL-1079 - Moved registration views into registrationview package
