+++ app/controllers/GrsRegistrationFailedController.scala
@@ -21,7 +21,7 @@ import controllers.actions.IdentifierAction
 import play.api.i18n.I18nSupport
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendBaseController
-import views.html.{RegistrationFailedNfmView, RegistrationFailedRfmView, RegistrationFailedUpeView}
+import views.html.registrationview.{RegistrationFailedNfmView, RegistrationFailedRfmView, RegistrationFailedUpeView}
 
 import javax.inject.{Inject, Named}
