+++ app/controllers/RegistrationWaitingRoomController.scala
@@ -24,7 +24,7 @@ import play.api.Logging
 import play.api.i18n.I18nSupport
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendBaseController
-import views.html.RegistrationWaitingRoomView
+import views.html.registrationview.RegistrationWaitingRoomView
 
 import javax.inject.Inject
