+++ app/controllers/eligibility/RegisteringNfmForThisGroupController.scala
@@ -41,7 +41,7 @@ import play.api.i18n.I18nSupport
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import repositories.SessionRepository
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendBaseController
-import views.html.RegisteringNfmForThisGroupView
+import views.html.registrationview.RegisteringNfmForThisGroupView
 
 import javax.inject.Inject
 import scala.concurrent.{ExecutionContext, Future}
