+++ app/views/registrationview/RegistrationWaitingRoomView.scala.html
